chore: move resouce files to own folder
