--- GR.Laromedel.AccessPortal.Web/src/components/about.js
@@ -1,20 +1,0 @@
-import React from 'react';
-
-export default function About() {
-  return (
-    <>
-      <h1 className="text-xl font-bold mb-4">Om Åtkomstportalen</h1>
-      <p>
-        Åtkomstportalen är skapad av{' '}
-        <a className="underline" href="https://goteborgsregionen.se">
-          Göteborgsregionen
-        </a>{' '}
-        för att göra den enkelt för elever och lärare att använda digitala läromedel.
-      </p>
-      <p>
-        Just nu är portalen under utveckling och den kommer förbättras och få flera funktioner
-        framöver.
-      </p>
-    </>
-  );
-}
