--- GR.Laromedel.AccessPortal.Web/src/components/privacy.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-export default function Privacy() {
-  return (
-    <>
-      <h1 className="text-xl font-bold mb-4">Cookies och persondata</h1>
-      <p>
-        Åtkomstportalen använder varken cookies eller persondata på något sätt för tillfället. Det
-        kommer däremot att ändras när man faktiskt kan logga in och göra saker. Då kommer det stå
-        något mer intressant här.
-      </p>
-    </>
-  );
-}
