+++ GR.Laromedel.AccessPortal.Web/src/components/resource/resource-list.js
