+++ GR.Laromedel.AccessPortal.Web/src/components/resource/resource.js
